started working on parcellations in native space
