+++ src/freesurfer_analyses/manager.py
@@ -3,6 +3,10 @@ import logging
 from pathlib import Path
 from typing import Union
 
+from brain_parts.parcellation.parcellations import (
+    Parcellation as parcellation_manager,
+)
+
 from freesurfer_analyses.utils.utils import LOGGER_CONFIG
 from freesurfer_analyses.utils.utils import collect_subjects
 from freesurfer_analyses.utils.utils import validate_instantiation
@@ -12,6 +16,10 @@ class FreesurferManager:
     BIDS_FILTERS = {"T1w": {"ceagent": "corrected"}}
     LOGGER_FILE = "freesurfer_analyses-{timestamp}.log"
 
+    #: Hemispheres
+    HEMISPHERES_LABELS = ["lh", "rh"]
+    SUBCORTICAL_LABELS = ["subcortex"]
+
     def __init__(
         self,
         base_dir: Path,
@@ -20,6 +28,7 @@ class FreesurferManager:
     ) -> None:
         self.data_grabber = validate_instantiation(self, base_dir)
         self.subjects = collect_subjects(self, participant_labels)
+        self.parcellation_manager = parcellation_manager()
         self.initiate_logging(logging_destination)
 
     def initiate_logging(self, logging_destination: Path = None) -> None:
@@ -42,6 +51,33 @@ class FreesurferManager:
             filename=str(
                 logging_destination
                 / self.LOGGER_FILE.format(timestamp=timestamp)
-            ),  # noqa
+            ),
             **LOGGER_CONFIG,
         )
+
+    def validate_session(
+        self, participant_label: str, session: Union[str, list] = None
+    ) -> list:
+        """
+        Validates session's input type (must be list)
+
+        Parameters
+        ----------
+        participant_label : str
+            Specific participants' labels
+        session : Union[str, list], optional
+            Specific session(s)' labels, by default None
+
+        Returns
+        -------
+        list
+            Either specified or available session(s)' labels
+        """
+        if session:
+            if isinstance(session, str):
+                sessions = [session]
+            elif isinstance(session, list):
+                sessions = session
+        else:
+            sessions = self.subjects.get(participant_label)
+        return sessions
